+++ scripts/_init.mjs
@@ -25,16 +25,6 @@ Hooks.once('ready', async () => {
 	
     Hooks.callAll('moduleCreditsReady');
 });
-
-Hooks.once('libChangelogsReady', async () => {
-	console.log('registerConflict');
-	libChangelogs.registerConflict(
-		'module-credits',
-		'lib-changelogs',
-		'Provides similar functionality.',
-		'minor'
-	)
-});
 	
 // FOUNDRY HOOKS -> REDNER SETTINGS CONFIG
 Hooks.on("renderSettingsConfig", (app, html) => {
@@ -46,8 +36,45 @@ Hooks.on("renderModuleManagement", (app, html) => {
 	ModuleCredits.renderModuleManagement(app, html);
 
 	ModuleCredits.conflicts.forEach((conflict, index) => {
-		console.log(conflict, `#module-management #module-list .package[data-module-name="${conflict.moduleID}"]`)
-		$(`#module-management #module-list .package[data-module-name="${conflict.moduleID}"]`).addClass(`module-credits-conflict-${conflict.status}`);
-		$(`#module-management #module-list .package[data-module-name="${conflict.conflictingModuleID}"]`).addClass(`module-credits-conflict-${conflict.status}`);
+		let $package = $(`#module-management #module-list .package[data-module-name="${conflict.moduleID}"]`);
+		$package.find('.package-title input').after('<i class="module-credits-conflict fas fa-exclamation-triangle"></i>');
+
+		let tooltip = `## Conflicts with ${game.modules.get(conflict.conflictingModuleID).data.title}
+${conflict.content}`;
+		$package.find('.module-credits-conflict').data('conflict', tooltip);
+
+		$package = $(`#module-management #module-list .package[data-module-name="${conflict.conflictingModuleID}"]`);
+		$package.find('.package-title input').after('<i class="module-credits-conflict fas fa-exclamation-triangle"></i>');
+
+		tooltip = `## Conflicts with ${game.modules.get(conflict.moduleID).data.title}
+${conflict.content}`;
+		$package.find('.module-credits-conflict').data('conflict', tooltip);
+	});
+
+	let popperInstance = null;
+	$('.module-credits-conflict').on('mouseenter focus', (event) => {
+		let $self = $(event.target);
+		let $tooltip = $(`<div id="${MODULE.name}-tooltip" role="tooltip">
+				<div id="${MODULE.name}-arrow" data-popper-arrow></div>
+			</div>`);
+		$tooltip.prepend(MODULE.markup($self.data('conflict')));
+		$self.before($tooltip);
+
+		popperInstance = Popper.createPopper($self[0], $tooltip[0], {
+			modifiers: [{
+				name: 'offset',
+				options: {
+					offset: [0, 8],
+				}
+			}]
+		});
+
+		popperInstance.update();
 	})
+
+
+	$('.module-credits-conflict').on('mouseleave blur', (event) => {
+		popperInstance.destroy();
+		$(`#${MODULE.name}-tooltip`).remove();
+	});
 });