+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-# Version 0.0.4.5 - Changelogs & Conflicts Support
+# Version 0.0.5.0 - The Story of Conflicts and More.
 - With permission from theripper93 I've added a wrapper to support modules who use his module in Module Credits
 - Fixed Author Popover being under elements below it
 - Added the ability to define a changelog in code with `game.modules.get('module-credits').api.registerChangelog({...})`
 - Added the ability to define a conflict with `game.modules.get('module-credits').api.defineConflict({...})`
 - Added Global `conflicts.json` file that is read from github to allow conflicts to be defined before modules are enabled
+- Added Support for Manifest Plus `conflicts` section
+- Added Ko-fi and reddit support for author credits
 
 # Version 0.0.4.0 - Plus more Options
 - Updated to latest version of Foundry.